I've finished single arrays lesson.
Next step is to start with matrix. The last exercise was something of a
challange, but it made me wonder about when will the MakeFile be
useful? I'm already searching somewhere I can learn, YouTube will help
me... Again.
I also did a .Session for examples and for exercises. It makes it simple
to track the content in VIM, a good text editor, the best IDE.