Reached a new chapter
So close, I'm so close to learn how to manage pointers.
I know they're pretty hard to understand, but there are lots of
exercises in the book. Actually, I began reading it because of these
exercises.
I've added the newer examples and exercises to my Session in vim. Today
it may seen useless, but tomorrow it will help me to revisit it all