A Few per day...
I'm finishing chapter 7, where we're seeing the linked list data
structure. It kind of take me to a dynamic feature at a static type
language, which means that we're gaining higher level powers. Later I'll
be playing with files, and I'll finally end the 'basics' of some of the
most important programming languages.