Just two more exercises
I'm not sure why I do commits in english since this repo is almost
entirely in portuguese. But I'm almost at the end of the chapter 7, this
time for real. chapter eight is a small one, and after it I'll be able
to go after more advanced stuff about proggraming. Let's do it