Ready to start a Space Invaders text based game.
I'm not sure if it's a good idea, since it can get very long. But
anyway, it's ready. I'm thinking about the other projects I want to
make, and if it's not a waste of time doing this project. But honestly,
to practice some programming is a nice thing to do. I'll slow down
everything, and keep doing everything as the same, but at a slower pace.
Here we go.