Now with a new structure I can advance
It was really hard to break things down to start building things up, but
with the new workflow it also feels better to code. Many things can be
changed and I can see their impacts, so... It's better.
The main difference is that I can test some pieces of code and run
anyway.
And the project structure helps me to hide what is finished. With this I
can just go straight.