Ready to go to the next step.
The next step is to implement a function that uses the random utilities.
I'm not very confortable with it, so I've been working on the other
functions that can be tailored to get better behavior, or that can be
optimized (the case with imprimeMatriz.c). Now I'll wait a bit 'til I
get back to work on this project.