Now the enemies can shoot you
I've added a empty file at the bin and obj directories because they
won't get added to the commit if they're empty. And now one more
function exists, and the enemies can shoot you. Nothing exactly new was
learned in this project, but at the end I'll refactor everything, so the
code will be smaller, and will fit better the necessities of a very
small memory, in the way it was when we needed to shrink every bit we
could. I hope this project to not to last longer.