Corrected some bugs
In main, joga, and moveNaves, there were some bugs. And also, I did a
little review of the code. There are only two functions missing, and
they are pretty similar. They're also similar to MoveNaves, which is
quite big. I'm thinking about to scatter the function so it became more
general and I'll be able to use one function to get the work done